+++ lgrez/bdd/model_joueurs.py
@@ -4,6 +4,8 @@ Déclaration de toutes les tables et leurs colonnes
 
 """
 
+import datetime
+
 import sqlalchemy
 from sqlalchemy.ext.hybrid import hybrid_property
 
@@ -58,6 +60,8 @@ class Joueur(base.TableBase):
         doc="Actions pour ce joueur")
     candidharos = autodoc_OneToMany("CandidHaro", back_populates="joueur",
         doc="Candidatures et haros de/contre ce joueur")
+    bouderies = autodoc_OneToMany("Bouderie", back_populates="joueur",
+        doc="Appartenances aux boudoirs de ce joueur")
     ciblages = autodoc_DynamicOneToMany("Ciblage", back_populates="joueur",
         doc="Ciblages prenant ce joueur pour cible")
 
@@ -108,6 +112,11 @@ class Joueur(base.TableBase):
         Élimine aussi les actions de vote (sans base).
         """
         return [ac for ac in self.actions if ac.active and not ac.vote]
+
+    @property
+    def boudoirs(self):
+        """Sequence[.Boudoir]: Boudoirs où est ce jour (read-only)"""
+        return [boud.boudoir for boud in self.bouderies]
 
     @hybrid_property
     def est_vivant(self):
@@ -205,3 +214,143 @@ class CandidHaro(base.TableBase):
     def __repr__(self):
         """Return repr(self)."""
         return f"<CandidHaro #{self.id} ({self.joueur}/{self.type})>"
+
+
+class Boudoir(base.TableBase):
+    """Table de données des boudoirs sur le serveur.
+
+    Les instances de cette classe sont crées, modifiées et supprimées par
+    :meth:`\!boudoir <.chans.GestionChans.GestionChans.boudoir.callback>`.
+    """
+    chan_id = autodoc_Column(sqlalchemy.BigInteger(), primary_key=True,
+        autoincrement=False, doc="ID Discord du salon")
+
+    nom = autodoc_Column(sqlalchemy.String(32), nullable=False,
+        doc="Nom du boudoir (demandé à la création)")
+    ts_created = autodoc_Column(sqlalchemy.DateTime(), nullable=False,
+        doc="Timestamp de la création")
+
+    # One-to-manys
+    bouderies = autodoc_OneToMany("Bouderie", back_populates="boudoir",
+        doc="Appartenances à ce boudoir")
+
+    def __repr__(self):
+        """Return repr(self)."""
+        return f"<Boudoir #{self.chan_id} ({self.nom})>"
+
+    @property
+    def chan(self):
+        """discord.TextChannel: Salon Discord correspondant à ce boudoir.
+
+        Raises:
+            ValueError: pas de membre correspondant
+            ~ready_check.NotReadyError: bot non connecté
+                (:attr:`.config.guild` vaut ``None``)
+        """
+        result = config.guild.get_channel(self.chan_id)
+        if not result:
+            raise ValueError(f"Boudoir.chan : pas de chan pour {self} !")
+
+        return result
+
+    @property
+    def joueurs(self):
+        """Sequence[.Joueur]: Joueurs présents dans ce boudoir (read-only)"""
+        return [boud.joueur for boud in self.bouderies]
+
+    @property
+    def gerant(self):
+        """.bdd.Joueur: Membre du boudoir ayant les droits de gestion.
+
+        Raises:
+            ValueError: pas de membre avec les droits de gestion
+        """
+        try:
+            return next(boud.joueur for boud in self.bouderies if boud.gerant)
+        except StopIteration:
+            raise ValueError(f"Pas de membre gérant le boudoir *{self.name}*")
+
+    async def add_joueur(self, joueur, gerant=False):
+        """Ajoute un joueur sur le boudoir.
+
+        Crée la :class:`.Bouderie` correspondante et modifie les
+        permissions du salon.
+
+        Args:
+            joueur (.Joueur): Le joueur à ajouter.
+            gerant (bool): Si le joueur doit être ajouté avec les
+                permissions de gérant.
+        """
+        now = datetime.datetime.now()
+        Bouderie(boudoir=self, joueur=joueur, gerant=gerant,
+                 ts_added=now, ts_promu=now if gerant else None).add()
+        await self.chan.set_permissions(joueur.member, read_messages=True)
+
+    async def remove_joueur(self, joueur):
+        """Retire un joueur du boudoir.
+
+        Supprime la :class:`.Bouderie` correspondante et modifie les
+        permissions du salon.
+
+        Args:
+            joueur (.Joueur): Le joueur à ajouter.
+        """
+        Bouderie.query.filter_by(boudoir=self, joueur=joueur).one().delete()
+        await self.chan.set_permissions(joueur.member, overwrite=None)
+
+    @classmethod
+    def from_channel(cls, channel):
+        """Récupère le Boudoir (instance de BDD) lié à un salon Discord.
+
+        Args:
+            channel (discord.TextChannel): le salon concerné
+
+        Returns:
+            .Boudoir: Le salon correspondant.
+
+        Raises:
+            ValueError: boudoir introuvable en base
+            ~ready_check.NotReadyError: session non initialisée
+                (:attr:`.config.session` vaut ``None``)
+        """
+        boudoir = cls.query.get(channel.id)
+        if not boudoir:
+            raise ValueError("Boudoir.from_channel : "
+                             f"pas de boudoir en base pour `{channel.mention}` !")
+
+        return boudoir
+
+
+class Bouderie(base.TableBase):
+    """Table de données des appartenances aux boudoirs du serveur.
+
+    Table d'association entre :class:`.Joueur` et :class:`.Boudoir`.
+
+    Les instances de cette classe sont crées, modifiées et supprimées par
+    :meth:`\!boudoir <.chans.GestionChans.GestionChans.boudoir.callback>`.
+    """
+    id = autodoc_Column(sqlalchemy.Integer(), primary_key=True,
+        doc="Identifiant unique de la bouderie, sans signification")
+
+    _joueur_id = sqlalchemy.Column(sqlalchemy.ForeignKey("joueurs.discord_id"),
+        nullable=False)
+    joueur = autodoc_ManyToOne("Joueur", back_populates="bouderies",
+        doc="Joueur concerné")
+
+    _boudoir_id = sqlalchemy.Column(sqlalchemy.ForeignKey("boudoirs.chan_id"),
+        nullable=False)
+    boudoir = autodoc_ManyToOne("Boudoir", back_populates="bouderies",
+        doc="Boudoir concerné")
+
+    gerant = autodoc_Column(sqlalchemy.Boolean(), nullable=False,
+        default=False,
+        doc="Si le joueur a les droits de gestion du boudoir")
+
+    ts_added = autodoc_Column(sqlalchemy.DateTime(), nullable=False,
+        doc="Timestamp de l'ajout du joueur au boudoir")
+    ts_promu = autodoc_Column(sqlalchemy.DateTime(),
+        doc="Timestamp de la promotion en gérant, le cas échéant")
+
+    def __repr__(self):
+        """Return repr(self)."""
+        return f"<Bouderie #{self.id} ({self.joueur}/{self.boudoir})>"